pyproject.toml
[project]
name = "openai-agent-sdk"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "SarfrazAziz786", email = "162900723+SarfrazAziz786@users.noreply.github.com" }
]
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.1.0",
    "python-decouple>=3.8",
    "rich>=14.0.0",
]

[project.scripts]
openai-agent-sdk = "openai_agent_sdk:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
